PIM-5375: Move some classes which was in TODO in upgrade.md
